change index.py

+++ index.py
@@ -9,19 +9,24 @@ def output_all_records(cars):
                 
 def output_for_recording_by_id(cars):
         num=input("Введите номер записи по которой вы хотите сделать вывод информации:\n")
-        with open('cars.json', 'r', encoding='utf-8') as file:
-            data = json.load(file)
-            found = False
-            for car in data:
-                if car['id'] == num:
-                    print(f"\n=============== Найдено ===============")
-                    print(f"{car['id']} >> Модель: {car['name']}, Производитель: {car['manufacturer']}")
-                    print(f"Заправляется бензином: {car['is_petrol']}, Объем бака: {car['tank_volume']}")
-                    found = True
-                    break
-            if not found:
-                print("\n=============== Не найдено ===============")
+        if num.isdigit():
+            with open('cars.json', 'r', encoding='utf-8') as file:
+                data = json.load(file)
+                found = False
+                for car in data:
+                    if car['id'] == num:
+                        print(f"\n=============== Найдено ===============")
+                        print(f"{car['id']} >> Модель: {car['name']}, Производитель: {car['manufacturer']}")
+                        print(f"Заправляется бензином: {car['is_petrol']}, Объем бака: {car['tank_volume']}")
+                        found = True
+                        break
+                if not found:
+                    print("\n=============== Не найдено ===============")
 
+        else:
+            print("Введено не число")
+
+            
 def adding_a_record(cars):
     with open('cars.json', 'r+', encoding='utf-8') as file:
         data = json.load(file)
@@ -34,49 +39,55 @@ def adding_a_record(cars):
             new_id = max(int(car['id']) for car in data) + 1
 
         new_name = input("Введите название модели автомобиля: ")
-        new_manufacturer = input("Введите название производителя: ")
+        new_manufacturer = input("Введите название производителя: ")      
         new_is_petrol = input("Заправляется бензином? (True/False): ").strip().lower() == 'true'
-        new_tank_volume = int(input("Введите объем бака: "))
+        new_tank_volume = input("Введите объем бака: ")
+        if new_tank_volume.isdigit():
 
-        new_car = {
-            "id": str(new_id),  # Store ID as a string for consistency with the existing data
-            "name": new_name,
-            "manufacturer": new_manufacturer,
-            "is_petrol": new_is_petrol,
-            "tank_volume": new_tank_volume
-        }
+            new_car = {
+                "id": str(new_id),  # Store ID as a string for consistency with the existing data
+                "name": new_name,
+                "manufacturer": new_manufacturer,
+                "is_petrol": new_is_petrol,
+                "tank_volume": new_tank_volume
+            }
 
-        # Add the new car record
-        data.append(new_car)
+            # Add the new car record
+            data.append(new_car)
 
-        # Save updated data back to the file
-        file.seek(0)
-        json.dump(data, file, ensure_ascii=False, indent=4)
-        file.truncate()
-
-    print(f"Запись добавлена: {new_car}")
-
-    
-def delete_record(cars):
-    delete_id = input("Введите номер записи, которую вы хотите удалить: ")
-    with open('cars.json', 'r+', encoding='utf-8') as file:
-        data = json.load(file)
-        found = False
-        # Look for the car to delete
-        for car in data:
-            if car['id'] == delete_id:
-                data.remove(car)
-                found = True
-                break
-
-        if not found:
-            print("\n=============== Не найдено ===============")
-        else:
             # Save updated data back to the file
             file.seek(0)
             json.dump(data, file, ensure_ascii=False, indent=4)
             file.truncate()
-            print(f"\nЗапись с номером {delete_id} удалена.")
+            print(f"Запись добавлена: {new_car}")
+
+        else:
+            print("Введите правильный объем бака.Введено не число")
+
+    
+def delete_record(cars):
+    delete_id = input("Введите номер записи, которую вы хотите удалить: ")
+    if delete_id.isdigit():
+        with open('cars.json', 'r+', encoding='utf-8') as file:
+            data = json.load(file)
+            found = False
+            # Look for the car to delete
+            for car in data:
+                if car['id'] == delete_id:
+                    data.remove(car)
+                    found = True
+                    break
+
+            if not found:
+                print("\n=============== Не найдено ===============")
+            else:
+                # Save updated data back to the file
+                file.seek(0)
+                json.dump(data, file, ensure_ascii=False, indent=4)
+                file.truncate()
+                print(f"\nЗапись с номером {delete_id} удалена.")
+    else:
+        print("Введено не число")
 
 
 def main():
